+++ cvat-ui/src/containers/annotation-page/standard-workspace/objects-side-bar/objects-list.tsx
@@ -127,15 +127,27 @@ const componentShortcuts = {
         scope: ShortcutScope.OBJECTS_SIDEBAR,
     },
     TO_BACKGROUND: {
-        name: 'To background',
-        description: 'Put an active object "farther" from the user (decrease z axis value)',
+        name: 'Move to background',
+        description: 'Move an active object to the newly created background layer (decrease z-order value)',
         sequences: ['-', '_'],
         scope: ShortcutScope.OBJECTS_SIDEBAR,
     },
     TO_FOREGROUND: {
-        name: 'To foreground',
-        description: 'Put an active object "closer" to the user (increase z axis value)',
+        name: 'Move to foreground',
+        description: 'Move an active object to the newly created foreground layer (increase z-order value)',
         sequences: ['+', '='],
+        scope: ShortcutScope.OBJECTS_SIDEBAR,
+    },
+    TO_ONE_LAYER_BACKWARD: {
+        name: 'Move one layer backward',
+        description: 'Move an active object one layer backward (decrease z-order value)',
+        sequences: [],
+        scope: ShortcutScope.OBJECTS_SIDEBAR,
+    },
+    TO_ONE_LAYER_FORWARD: {
+        name: 'Move one layer forward',
+        description: 'Move an active object one layer forward (increase z-order value)',
+        sequences: [],
         scope: ShortcutScope.OBJECTS_SIDEBAR,
     },
     COPY_SHAPE: {
@@ -457,7 +469,7 @@ class ObjectsListContainer extends React.PureComponent<Props, State> {
             objectStates, sortedStatesID, statesOrdering, filteredStates,
         } = this.state;
 
-        const preventDefault = (event: KeyboardEvent | undefined): void => {
+        const preventDefault = (event?: KeyboardEvent): void => {
             if (event) {
                 event.preventDefault();
             }
@@ -481,11 +493,11 @@ class ObjectsListContainer extends React.PureComponent<Props, State> {
         };
 
         const handlers: Record<keyof typeof componentShortcuts, (event?: KeyboardEvent) => void> = {
-            SWITCH_ALL_LOCK: (event: KeyboardEvent | undefined) => {
+            SWITCH_ALL_LOCK: (event?: KeyboardEvent) => {
                 preventDefault(event);
                 this.lockAllStates(!statesLocked);
             },
-            SWITCH_LOCK: (event: KeyboardEvent | undefined) => {
+            SWITCH_LOCK: (event?: KeyboardEvent) => {
                 preventDefault(event);
                 const state = activatedState();
                 if (state) {
@@ -493,11 +505,11 @@ class ObjectsListContainer extends React.PureComponent<Props, State> {
                     updateAnnotations([state]);
                 }
             },
-            SWITCH_ALL_HIDDEN: (event: KeyboardEvent | undefined) => {
+            SWITCH_ALL_HIDDEN: (event?: KeyboardEvent) => {
                 preventDefault(event);
                 this.hideAllStates(!statesHidden);
             },
-            SWITCH_HIDDEN: (event: KeyboardEvent | undefined) => {
+            SWITCH_HIDDEN: (event?: KeyboardEvent) => {
                 preventDefault(event);
                 const state = activatedState();
                 const {
@@ -512,7 +524,7 @@ class ObjectsListContainer extends React.PureComponent<Props, State> {
                     updateAnnotations([state]);
                 }
             },
-            SWITCH_OCCLUDED: (event: KeyboardEvent | undefined) => {
+            SWITCH_OCCLUDED: (event?: KeyboardEvent) => {
                 preventDefault(event);
                 const state = activatedState();
                 if (state && state.objectType !== ObjectType.TAG) {
@@ -520,7 +532,7 @@ class ObjectsListContainer extends React.PureComponent<Props, State> {
                     updateAnnotations([state]);
                 }
             },
-            SWITCH_PINNED: (event: KeyboardEvent | undefined) => {
+            SWITCH_PINNED: (event?: KeyboardEvent) => {
                 preventDefault(event);
                 const state = activatedState(true);
                 if (state) {
@@ -528,7 +540,7 @@ class ObjectsListContainer extends React.PureComponent<Props, State> {
                     updateAnnotations([state]);
                 }
             },
-            SWITCH_KEYFRAME: (event: KeyboardEvent | undefined) => {
+            SWITCH_KEYFRAME: (event?: KeyboardEvent) => {
                 preventDefault(event);
                 const state = activatedState();
                 if (state && state.objectType === ObjectType.TRACK) {
@@ -539,7 +551,7 @@ class ObjectsListContainer extends React.PureComponent<Props, State> {
                     }
                 }
             },
-            SWITCH_OUTSIDE: (event: KeyboardEvent | undefined) => {
+            SWITCH_OUTSIDE: (event?: KeyboardEvent) => {
                 preventDefault(event);
                 const state = activatedState();
                 if (state && (state.objectType === ObjectType.TRACK || state.parentID)) {
@@ -547,14 +559,14 @@ class ObjectsListContainer extends React.PureComponent<Props, State> {
                     updateAnnotations([state]);
                 }
             },
-            DELETE_OBJECT_STANDARD_WORKSPACE: (event: KeyboardEvent | undefined) => {
+            DELETE_OBJECT_STANDARD_WORKSPACE: (event?: KeyboardEvent) => {
                 preventDefault(event);
                 const state = activatedState(true);
                 if (state) {
                     removeObject(state, event ? event.shiftKey : false);
                 }
             },
-            CHANGE_OBJECT_COLOR: (event: KeyboardEvent | undefined) => {
+            CHANGE_OBJECT_COLOR: (event?: KeyboardEvent) => {
                 preventDefault(event);
                 const state = activatedState();
                 if (state) {
@@ -571,7 +583,7 @@ class ObjectsListContainer extends React.PureComponent<Props, State> {
                     }
                 }
             },
-            TO_BACKGROUND: (event: KeyboardEvent | undefined) => {
+            TO_BACKGROUND: (event?: KeyboardEvent) => {
                 preventDefault(event);
                 const state = activatedState(true);
                 if (state && state.objectType !== ObjectType.TAG) {
@@ -579,11 +591,27 @@ class ObjectsListContainer extends React.PureComponent<Props, State> {
                     updateAnnotations([state]);
                 }
             },
-            TO_FOREGROUND: (event: KeyboardEvent | undefined) => {
+            TO_FOREGROUND: (event?: KeyboardEvent) => {
                 preventDefault(event);
                 const state = activatedState(true);
                 if (state && state.objectType !== ObjectType.TAG) {
                     state.zOrder = maxZLayer + 1;
+                    updateAnnotations([state]);
+                }
+            },
+            TO_ONE_LAYER_BACKWARD: (event?: KeyboardEvent) => {
+                preventDefault(event);
+                const state = activatedState(true);
+                if (state && state.objectType !== ObjectType.TAG) {
+                    state.zOrder -= 1;
+                    updateAnnotations([state]);
+                }
+            },
+            TO_ONE_LAYER_FORWARD: (event?: KeyboardEvent) => {
+                preventDefault(event);
+                const state = activatedState(true);
+                if (state && state.objectType !== ObjectType.TAG) {
+                    state.zOrder += 1;
                     updateAnnotations([state]);
                 }
             },
@@ -601,14 +629,14 @@ class ObjectsListContainer extends React.PureComponent<Props, State> {
                     openAnnotationsActionModal();
                 }
             },
-            PROPAGATE_OBJECT: (event: KeyboardEvent | undefined) => {
+            PROPAGATE_OBJECT: (event?: KeyboardEvent) => {
                 preventDefault(event);
                 const state = activatedState();
                 if (state) {
                     switchPropagateVisibility(true);
                 }
             },
-            NEXT_KEY_FRAME: (event: KeyboardEvent | undefined) => {
+            NEXT_KEY_FRAME: (event?: KeyboardEvent) => {
                 preventDefault(event);
                 const state = activatedState();
                 if (state && state.keyframes) {
@@ -618,7 +646,7 @@ class ObjectsListContainer extends React.PureComponent<Props, State> {
                     }
                 }
             },
-            PREV_KEY_FRAME: (event: KeyboardEvent | undefined) => {
+            PREV_KEY_FRAME: (event?: KeyboardEvent) => {
                 preventDefault(event);
                 const state = activatedState();
                 if (state && state.keyframes) {