+++ cvat/apps/dataset_manager/bindings.py
@@ -1589,13 +1589,9 @@ class MediaProvider3D(MediaProvider):
     def get_media_for_frame(self, source_id: int, frame_id: int, **image_kwargs) -> dm.PointCloud:
         source = self._sources[source_id]
 
-        upload_dir = source.db_task.data.get_upload_dirname()
         point_cloud_path = image_kwargs['path']
 
-        related_image_paths = [
-            osp.relpath(str(ri_path), upload_dir)
-            for ri_path in self._ri_per_source[source_id].get(frame_id, [])
-        ]
+        related_image_paths = self._ri_per_source[source_id].get(frame_id, [])
 
         def get_pcd_bytes():
             self._load_source(source_id, source)
@@ -1629,7 +1625,7 @@ class MediaProvider3D(MediaProvider):
 
             frame_related_images = {
                 ri_path: Path(ri_realpath).read_bytes()
-                for _, (ri_realpath, ri_path, _) in cache.read_raw_context_images(
+                for _, (ri_realpath, ri_path) in cache.read_raw_context_images(
                     self._sources[self._current_source_id].db_task.data,
                     frame_ids=[frame_id],
                     truncate_common_filename_prefix=False,